requirements.txt
asgiref==3.2.7
astroid==2.4.1
boto==2.49.0
boto3==1.14.20
botocore==1.17.20
certifi==2020.4.5.1
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
dj-database-url==0.5.0
Django==3.0.7
django-allauth==0.42.0
django-crispy-forms==1.9.1
django-excel==0.0.10
django-filter==2.3.0
django-heroku==0.3.1
django-import-export==2.2.0
django-storages==1.9.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
docutils==0.15.2
et-xmlfile==1.0.1
gunicorn==20.0.4
httpie==2.2.0
idna==2.9
isort==4.3.21
jdcal==1.4.1
jmespath==0.10.0
lazy-object-proxy==1.4.3
lml==0.0.9
Markdown==3.2.2
MarkupPy==1.14
mccabe==0.6.1
mysqlclient==2.0.1
numpy==1.18.4
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.3
pandas==1.0.4
Pillow==7.2.0
psycopg2==2.8.5
pyexcel==0.6.2
pyexcel-io==0.5.20
pyexcel-webio==0.1.4
Pygments==2.6.1
PyJWT==1.7.1
pyperclip==1.8.0
python-dateutil==2.8.1
python-http-client==3.2.7
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.3.1
pyzmail36==1.0.4
requests==2.23.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
sendgrid==6.3.1
six==1.15.0
sql==0.4.0
sqlparse==0.3.1
tablib==2.0.0
texttable==1.6.2
toml==0.10.1
urllib3==1.25.9
whitenoise==5.1.0
wrapt==1.12.1
xlrd==1.2.0
xlwt==1.3.0
